+++ backend/manager/pom.xml
@@ -8,7 +8,7 @@
         <version>3.3.4</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
-<!--    <groupId>com.example</groupId> -->
+    <groupId>org.github.babkiniaa.scas</groupId>
     <artifactId>demo</artifactId>
 <!--    <version>0.0.1-SNAPSHOT</version> -->
     <name>demo</name>
