--- backend/manager/src/main/java/com/example/demo/entity/Role.java
@@ -1,28 +1,0 @@
-package com.example.demo.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- * Сущность, представляющая роль пользователя в системе.
- * Каждая роль используется для управления правами доступа пользователей.
- */
-@Entity
-@Setter
-@Getter
-@Table
-public class Role {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column
-  private String name;
-}
