--- backend/manager/src/main/java/com/example/demo/validatros/Password.java
@@ -1,17 +1,0 @@
-package com.example.demo.validatros;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.*;
-
-
-@Documented
-@Target(ElementType.FIELD)
-@Constraint(validatedBy = PasswordValidator.class)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Password {
-    String message() default "The password must contain uppercase and lowercase characters, numbers, and must be at least 8 characters long";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-}
