--- backend/manager/src/main/java/com/example/demo/validatros/PasswordValidator.java
@@ -1,19 +1,0 @@
-package com.example.demo.validatros;
-
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-
-public class PasswordValidator implements ConstraintValidator<Password, String> {
-
-    private static final String PASSWORD ="^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).{8,}$";
-
-    @Override
-    public void initialize(Password constraintAnnotation) {
-        ConstraintValidator.super.initialize(constraintAnnotation);
-    }
-
-    @Override
-    public boolean isValid(String s, ConstraintValidatorContext constraintValidatorContext) {
-        return s.matches(PASSWORD);
-    }
-}
