--- backend/manager/src/main/java/com/example/demo/validatros/UniqueEmail.java
@@ -1,16 +1,0 @@
-package com.example.demo.validatros;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.*;
-
-@Documented
-@Target({ElementType.PARAMETER, ElementType.FIELD})
-@Constraint(validatedBy = UniqueEmailValidator.class)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface UniqueEmail {
-    String message() default "The user with this email has already been registered";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-}
