--- backend/manager/src/main/java/com/example/demo/validatros/UniqueEmailValidator.java
@@ -1,22 +1,0 @@
-package com.example.demo.validatros;
-
-import com.example.demo.repository.UserRepository;
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-
-import org.springframework.beans.factory.annotation.Autowired;
-
-public class UniqueEmailValidator implements ConstraintValidator<UniqueEmail, String> {
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    public void initialize(UniqueEmail constraintAnnotation) {
-        ConstraintValidator.super.initialize(constraintAnnotation);
-    }
-
-    @Override
-    public boolean isValid(String email, ConstraintValidatorContext constraintValidatorContext) {
-        return userRepository.findByEmail(email).isEmpty();
-    }
-}
