--- backend/manager/src/main/java/com/example/demo/validatros/UniqueLoginValidator.java
@@ -1,22 +1,0 @@
-package com.example.demo.validatros;
-
-import com.example.demo.repository.UserRepository;
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-
-import org.springframework.beans.factory.annotation.Autowired;
-
-public class UniqueLoginValidator implements ConstraintValidator<UniqueLogin, String> {
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    public void initialize(UniqueLogin constraintAnnotation) {
-        ConstraintValidator.super.initialize(constraintAnnotation);
-    }
-
-    @Override
-    public boolean isValid(String login, ConstraintValidatorContext constraintValidatorContext) {
-        return userRepository.findByLogin(login).isEmpty();
-    }
-}
