+++ backend/manager/src/main/java/org/github/babkiniaa/scas/entity/Token.java
@@ -1,0 +1,40 @@
+package org.github.babkiniaa.scas.entity;
+
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import java.time.LocalDateTime;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+/**
+ * Сущность для хранения токенов, используемых для верификации пользователей.
+ * Токены применяются, например, для подтверждения регистрации по электронной почте.
+ */
+@Table
+@Getter
+@Setter
+@Entity(name = "token")
+@RequiredArgsConstructor
+public class Token {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true)
+    private String verificationToken;
+
+    @OneToOne(targetEntity = User.class, fetch = FetchType.EAGER)
+    @JoinColumn(nullable = false, name = "user_id")
+    private User user;
+
+    private LocalDateTime expiryDate;
+}
