+++ backend/manager/src/main/java/org/github/babkiniaa/scas/entity/User.java
@@ -1,0 +1,53 @@
+package org.github.babkiniaa.scas.entity;
+
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.Table;
+import java.util.HashSet;
+import java.util.Set;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+/**
+ * Сущность, представляющая пользователя в системе.
+ * Содержит информацию о пользователе, его ролях и состоянии верификации.
+ */
+@Table
+@Getter
+@Setter
+@Entity(name = "users")
+@RequiredArgsConstructor
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column
+    private String email;
+
+    @Column
+    private String login;
+
+    @Column
+    private String password;
+
+    @ManyToMany(fetch = FetchType.EAGER)
+    @JoinTable(name = "user_roles", joinColumns = @JoinColumn(name = "user_id"),
+            inverseJoinColumns = @JoinColumn(name = "role_id"))
+    private Set<Role> roles = new HashSet<>();
+
+    @Column
+    private String verificationCode;
+
+    @Column
+    private boolean isEnable = false;
+}
