+++ backend/manager/src/main/java/org/github/babkiniaa/scas/mappers/UserMapper.java
@@ -1,0 +1,18 @@
+package org.github.babkiniaa.scas.mappers;
+
+
+import org.github.babkiniaa.scas.dto.RegistrationDto;
+import org.github.babkiniaa.scas.entity.User;
+import org.mapstruct.Mapper;
+
+/**
+ * Интерфейс для преобразования между сущностями {@link User} и
+ * объектами передачи данных {@link RegistrationDto}.
+ * Использует MapStruct для автоматической генерации реализации маппинга.
+ */
+@Mapper(componentModel = "spring")
+public interface UserMapper {
+
+    User toEntity(RegistrationDto userRegistrationDto);
+
+}
