+++ backend/manager/src/main/java/org/github/babkiniaa/scas/validatros/Password.java
@@ -1,0 +1,17 @@
+package org.github.babkiniaa.scas.validatros;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+import java.lang.annotation.*;
+
+
+@Documented
+@Target(ElementType.FIELD)
+@Constraint(validatedBy = PasswordValidator.class)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Password {
+    String message() default "The password must contain uppercase and lowercase characters, numbers, and must be at least 8 characters long";
+    Class<?>[] groups() default {};
+    Class<? extends Payload>[] payload() default {};
+}
