+++ backend/manager/src/main/java/org/github/babkiniaa/scas/validatros/PasswordValidator.java
@@ -1,0 +1,19 @@
+package org.github.babkiniaa.scas.validatros;
+
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+
+public class PasswordValidator implements ConstraintValidator<Password, String> {
+
+    private static final String PASSWORD ="^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).{8,}$";
+
+    @Override
+    public void initialize(Password constraintAnnotation) {
+        ConstraintValidator.super.initialize(constraintAnnotation);
+    }
+
+    @Override
+    public boolean isValid(String s, ConstraintValidatorContext constraintValidatorContext) {
+        return s.matches(PASSWORD);
+    }
+}
